# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/advisors/{advisorId}/groups/{groupId}": {
      "delete": {
        "operationId": "AdvisorsController_releaseTeam",
        "parameters": [
          {
            "name": "advisorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Advisors"
        ]
      }
    },
...
    "/api/v1/groups/{groupId}/final-grade": {
      "get": {
        "operationId": "getGroupFinalGrade",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupFinalGradeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "No final grade calculated yet"
          },
          "500": {
            "description": "Unexpected internal failure"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get stored final grade for a group",
        "tags": [
          "Grades"
        ]
      }
    },
    "/api/v1/students/{studentId}/final-grade": {
      "get": {
        "operationId": "getStudentFinalGrade",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentFinalGradeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "No final grade calculated yet"
          },
          "500": {
            "description": "Unexpected internal failure"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get stored final grade for a student",
        "tags": [
          "Grades"
        ]
      }
    },
    "/api/v1/groups/{groupId}/grade-history": {
      "get": {
        "operationId": "getGradeHistory",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedGradeHistoryDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "No grade history found"
          },
          "500": {
            "description": "Unexpected internal failure"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get grade history snapshots for a group",
        "tags": [
          "Grades"
        ]
      }
    },
    "/api/v1/groups/{groupId}/calculate": {
      "post": {
        "operationId": "calculateGrade",
        "parameters": [
          {
            "name": "groupId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalculateGradeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GradeCalculationResultDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "409": {
            "description": "Final grade already exists and force=false"
          },
          "422": {
            "description": "Missing preconditions (evaluations not complete, config missing, etc)"
          },
          "500": {
            "description": "Unexpected internal failure"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Calculate and persist final grades for a group",
        "tags": [
          "Grades"
        ]
      }
    },
...
  },
  "info": {
    "title": "Senior App API",
    "description": "OpenAPI spec generated from backend routes",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "StudentFinalGradeDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "format": "uuid",
            "description": "Student identifier for the stored final grade"
          },
          "groupId": {
            "type": "string",
            "format": "uuid",
            "description": "Group identifier associated with this grade"
          },
          "individualAllowanceRatio": {
            "type": "number",
            "example": 0.9,
            "description": "Completed-to-target story point allowance ratio"
          },
          "finalGrade": {
            "type": "number",
            "example": 70.56,
            "description": "Stored individual final grade for the student"
          },
          "calculatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the final grade was calculated"
          }
        },
        "required": [
          "studentId",
          "groupId",
          "individualAllowanceRatio",
          "finalGrade",
          "calculatedAt"
        ]
      },
      "GroupFinalGradeDto": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string",
            "format": "uuid",
            "description": "Group identifier for the stored final grade"
          },
          "teamGrade": {
            "type": "number",
            "example": 78.4,
            "description": "Stored final team grade from D4"
          },
          "individualGrades": {
            "description": "Individual stored grades for students in the group",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentFinalGradeDto"
            }
          },
          "calculatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when this final grade was calculated"
          }
        },
        "required": [
          "groupId",
          "teamGrade",
          "individualGrades",
          "calculatedAt"
        ]
      },
      "GradeHistoryEntryDto": {
        "type": "object",
        "properties": {
          "gradeChangeId": {
            "type": "string",
            "format": "uuid"
          },
          "groupId": {
            "type": "string",
            "format": "uuid"
          },
          "teamGrade": {
            "type": "number",
            "example": 78.4
          },
          "gradeComponents": {
            "type": "object"
          },
          "triggeredBy": {
            "type": "string",
            "format": "uuid"
          },
          "changedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "gradeChangeId",
          "groupId",
          "teamGrade",
          "gradeComponents",
          "triggeredBy",
          "changedAt"
        ]
      },
      "PaginatedGradeHistoryDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Grade history entries for the requested group page",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GradeHistoryEntryDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of grade history entries"
          },
          "page": {
            "type": "number",
            "description": "Current page number"
          },
          "limit": {
            "type": "number",
            "description": "Requested page size"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ]
      },
      "CalculateGradeDto": {
        "type": "object",
        "properties": {
          "force": {
            "type": "boolean",
            "description": "When false (default), returns 409 if a final grade already exists. When true, recalculates and overwrites D4; a new snapshot is always appended to D5.",
            "default": false
          }
        },
        "required": [
          "force"
        ]
      },
      "ScaledDeliverableGradeDto": {
        "type": "object",
        "properties": {
          "deliverableId": {
            "type": "string",
            "format": "uuid",
            "description": "Deliverable identifier"
          },
          "rawGrade": {
            "type": "number",
            "example": 80,
            "description": "Raw numeric grade from committee evaluation (100/80/60/50/0)"
          },
          "teamScalar": {
            "type": "number",
            "example": 0.8,
            "description": "Team scalar [0–1]: average of all sprint evaluation scores for sprints mapped to this deliverable, divided by 100"
          },
          "categoryWeight": {
            "type": "number",
            "example": 0.5,
            "description": "Category weight from D1 deliverable config"
          },
          "subWeight": {
            "type": "number",
            "example": 0.35,
            "description": "Sub-weight from D1 deliverable config"
          },
          "scaledGrade": {
            "type": "number",
            "example": 22.4,
            "description": "Scaled grade: rawGrade × teamScalar × categoryWeight × subWeight"
          }
        },
        "required": [
          "deliverableId",
          "rawGrade",
          "teamScalar",
          "categoryWeight",
          "subWeight",
          "scaledGrade"
        ]
      },
      "GradeCalculationResultDto": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "string",
            "format": "uuid",
            "description": "Group identifier"
          },
          "teamGrade": {
            "type": "number",
            "example": 74.88,
            "description": "Sum of all scaled deliverable grades, capped at 100"
          },
          "teamScalar": {
            "type": "number",
            "example": 0.815,
            "description": "Overall team scalar: average of per-deliverable scalars"
          },
          "scaledDeliverableGrades": {
            "description": "Breakdown of scaled grades per deliverable",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScaledDeliverableGradeDto"
            }
          },
          "individualGrades": {
            "description": "Final grades per student (teamGrade × individualAllowanceRatio)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentFinalGradeDto"
            }
          },
          "triggeredBy": {
            "type": "string",
            "format": "uuid",
            "description": "userId of the coordinator who triggered the calculation"
          },
          "calculatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of calculation"
          }
        },
        "required": [
          "groupId",
          "teamGrade",
          "teamScalar",
          "scaledDeliverableGrades",
          "individualGrades",
          "triggeredBy",
          "calculatedAt"
        ]
      },
...
    }
  }
}